Created a new component for ConnectionLinks
